pyproject.toml
[project]
name = "stuffswapper-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Hellhound Waffle",email = "gambitbrian7m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "fastapi[standard] (>=0.121.3,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "fastapi-mail (>=1.5.8,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
